# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/category/{categoryId}/courses": {
            "get": {
                "summary": "Retrieves all courses within a specific category",
                "operationId": "12538fc3f1ab8b446392b6a989b08f92",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "The ID of the category for which courses are to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "course_id": {
                                                "description": "ID of the course",
                                                "type": "integer"
                                            },
                                            "course_title": {
                                                "description": "Title of the course",
                                                "type": "string"
                                            },
                                            "course_tags": {
                                                "description": "Tags associated with the course",
                                                "type": "string"
                                            },
                                            "course_thumbnail": {
                                                "description": "URL of the course thumbnail image",
                                                "type": "string"
                                            },
                                            "author": {
                                                "description": "Name of the course author",
                                                "type": "string"
                                            },
                                            "category_name": {
                                                "description": "Name of the category to which the course belongs",
                                                "type": "string"
                                            },
                                            "course_price": {
                                                "description": "Price of the course",
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "number_of_lessons": {
                                                "description": "Total number of lessons in the course",
                                                "type": "integer"
                                            },
                                            "number_of_enrollments": {
                                                "description": "Total number of enrollments in the course",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No courses found for the category"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/categories/get": {
            "get": {
                "summary": "Retrieves all course categories",
                "operationId": "e824db2ae9802b65a1f62448e3101876",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of all categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the category",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Name of the category",
                                                "type": "string"
                                            },
                                            "description": {
                                                "description": "Description of the category",
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "description": "Creation date of the category",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Last update date of the category",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/course/{courseId}/enroll": {
            "post": {
                "summary": "Enrolls a user into a specified course",
                "operationId": "ef1e20b63dc372d99bef75031c19b874",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "The ID of the course to enroll in",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User ID needed to enroll in the course",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "The user's ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrollment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "description": "User's ID",
                                            "type": "integer"
                                        },
                                        "course_id": {
                                            "description": "Course ID",
                                            "type": "integer"
                                        },
                                        "enrollment_status": {
                                            "description": "Status of the enrollment",
                                            "type": "string"
                                        },
                                        "enrollment_date": {
                                            "description": "Date and time of enrollment",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or user does not exist"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}